add QueryBuilder for fetch logics.
